Some changes in project structure for web deploy. Fixed bug with postrgres database initialization.
