+++ api/src/main/resources/api.properties
@@ -1,5 +1,5 @@
-# Url to application api (without leading "/")
-application.api.url=http://localhost:8080
-
 # Path to images directory (with leading "/")
 application.images.path=#{systemEnvironment['FASHION_RESOURCES_ROOT']}/images/
+
+# Url to application api (without leading "/")
+application.api.url=http://localhost:8080
