--- persistence/src/main/java/com/bionicuniversity/edu/fashiontips/entity/converter/LocalDateTimeConverter.java
@@ -1,22 +1,0 @@
-package com.bionicuniversity.edu.fashiontips.entity.converter;
-
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
-
-/**
- * LocaleDateTime Converter
- */
-@Converter(autoApply = true)
-public class LocalDateTimeConverter implements AttributeConverter<LocalDateTime, Timestamp> {
-    @Override
-    public Timestamp convertToDatabaseColumn(LocalDateTime attribute) {
-        return Timestamp.valueOf(attribute);
-    }
-
-    @Override
-    public LocalDateTime convertToEntityAttribute(Timestamp dbData) {
-        return dbData.toLocalDateTime();
-    }
-}
