+++ persistence/src/main/resources/db/tomcat.properties
@@ -1,0 +1,5 @@
+database.init=false
+
+jpa.showSql=false
+hibernate.format_sql=true
+hibernate.use_sql_comments=true
