+++ persistence/src/main/resources/persistence.properties
@@ -1,0 +1,2 @@
+# Path to images directory (with leading "/")
+application.images.path=#{systemEnvironment['FASHION_RESOURCES_ROOT']}/images/
