+++ service/src/main/java/com/bionicuniversity/edu/fashiontips/service/impl/ImageServiceImpl.java
@@ -3,7 +3,6 @@ package com.bionicuniversity.edu.fashiontips.service.impl;
 import com.bionicuniversity.edu.fashiontips.dao.ImageDao;
 import com.bionicuniversity.edu.fashiontips.entity.Image;
 import com.bionicuniversity.edu.fashiontips.service.ImageService;
-import com.bionicuniversity.edu.fashiontips.service.util.ImageUtil;
 
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -24,7 +23,6 @@ public class ImageServiceImpl implements ImageService {
     @Override
     public Image save(Image image) throws IOException {
         imageDao.save(image);
-        ImageUtil.createUrlName(image);
         return image;
     }
 }
