+++ service/src/main/resources/spring/spring-service.xml
@@ -10,14 +10,4 @@
     <!--Create service beans-->
     <context:component-scan base-package="com.bionicuniversity.edu.fashiontips.service"/>
 
-    <!-- Initialize ImageUtil with application URL -->
-    <bean class="org.springframework.beans.factory.config.MethodInvokingFactoryBean">
-        <property name="staticMethod" value="com.bionicuniversity.edu.fashiontips.service.util.ImageUtil.setApiUrl"/>
-        <property name="arguments">
-            <list>
-                <value>${application.api.url}</value>
-            </list>
-        </property>
-    </bean>
-
 </beans>
